requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
openai-whisper
google-generativeai>=0.3.0
elevenlabs>=1.0.0
python-dotenv>=1.0.0
aiofiles>=23.0.0
